+++ stor/obs.py
@@ -74,6 +74,14 @@ class OBSPath(Path):
     copy = utils.copy
     copytree = utils.copytree
 
+    def dirname(self):
+        """
+        Return the directory name of self.
+        """
+        assert self.startswith(self.drive)
+        rest = self[len(self.drive):]
+        return self.path_class(self.drive + self.path_module.dirname(rest))
+
     def __repr__(self):
         return '%s("%s")' % (type(self).__name__, self)